Fix Priority 2 code review issues: ISBN normalization (strip separators), @Max on pageSize, LIKE wildcard escaping, COUNT(*) mapped to Long instead of Integer

+++ backend/src/main/java/pl/jaro/restapiworkshop/repository/BookRepository.java
@@ -20,6 +20,6 @@ public interface BookRepository {
 
     void delete(Long id, Long userId);
 
-    int countAll(Long userId);
+    long countAll(Long userId);
 
 }
